pyproject.toml
[tool.poetry]
name = "pyutils"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.11"
pillow = "^10.4.0"
win32-setctime = "^1.1.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
